Fix iOS 10 archive failure by spliting swift UI property wrapper into dedicated pod subspec This offer iOS 10 compatibility by importing pod 'Resolver/Core' or 'Resolver' for swift UI project
