+++ Sources/Resolver/SwiftUI/Resolver+SwiftUI.swift
@@ -1,0 +1,45 @@
+//
+//  Resolver+SwiftUI.swift
+//  Resolver
+//
+//  Created by bfrolicher on 31/12/2019.
+//
+#if os(iOS)
+import UIKit
+import SwiftUI
+#elseif os(macOS) || os(tvOS) || os(watchOS)
+import Foundation
+import SwiftUI
+#else
+import Foundation
+#endif
+
+
+#if swift(>=5.1)
+/// Immediate injection property wrapper for SwiftUI ObservableObjects. This wrapper is meant for use in SwiftUI Views and exposes
+/// bindable objects similar to that of SwiftUI @observedObject and @environmentObject.
+///
+/// Dependent service must be of type ObservableObject. Updating object state will trigger view update.
+///
+/// Wrapped dependent service is resolved immediately using Resolver.root upon struct initialization.
+///
+@available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
+@propertyWrapper
+public struct InjectedObject<Service>: DynamicProperty where Service: ObservableObject {
+    @ObservedObject private var service: Service
+    public init() {
+        self.service = Resolver.resolve(Service.self)
+    }
+    public init(name: String? = nil, container: Resolver? = nil) {
+        self.service = container?.resolve(Service.self, name: name) ?? Resolver.resolve(Service.self, name: name)
+    }
+    public var wrappedValue: Service {
+        get { return service }
+        mutating set { service = newValue }
+    }
+    public var projectedValue: ObservedObject<Service>.Wrapper {
+        return self.$service
+    }
+}
+
+#endif
